* 0407_PropogatePolicies_zach_mullen:
  STYLE: refs #0407. Fix trailing whitespace
  ENH: refs #0407. Add a test for policy management behavior
  ENH: refs #0407. Only admins should see permissions and delete links
  BUG: refs #0407. Require admin permission to delete a folder
  ENH: refs #0407. Rewrite existing permissions dialog controller
  ENH: refs #0407. Implement recursive policy copying in PolicyComponent
  ENH: refs #0407. Client side code and stub ctrlr for apply recursive policies
  ENH: refs #0407. Improve our dialog api to have no minimum height.

+++ core/public/js/common/common.browser.js
@@ -345,10 +345,16 @@ function createAction(node)
         {
         html+='<li><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/folder_add.png"/> <a onclick="createNewFolder('+element+');">'+json.browse.createFolder+'</a></li>';
         html+='<li><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/upload.png"/> <a rel="'+json.global.webroot+'/upload/simpleupload/?parent='+element+'" class="uploadInFolder">'+json.browse.uploadIn+'</a></li>';
-                  html+='<li><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/lock.png"/> <a type="folder" element="'+element+'" class="sharingLink">'+json.browse.share+'</a></li>';
         if(node.attr('deletable')!=undefined && node.attr('deletable')=='true')
           {
           html+='<li><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/edit.png"/> <a onclick="editFolder('+element+');">'+json.browse.edit+'</a></li>';
+          }
+        }
+      if(policy>=2)
+        {
+        html+='<li><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/lock.png"/> <a type="folder" element="'+element+'" class="sharingLink">'+json.browse.share+'</a></li>';
+        if(node.attr('deletable')!=undefined && node.attr('deletable')=='true')
+          {
           html+='<li><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/close.png"/> <a onclick="deleteFolder('+element+');">'+json.browse['delete']+'</a></li>';
           }
         }
@@ -357,7 +363,7 @@ function createAction(node)
       {
       html+='<li><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/view.png"/> <a href="'+json.global.webroot+'/item/'+element+'">'+json.browse.view+'</a></li>';
       html+='<li><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/download.png"/> <a href="'+json.global.webroot+'/download?items='+element+'">'+json.browse.downloadLatest+'</a></li>';
-      if(policy>=1)
+      if(policy>=2)
         {
         html+='<li><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/lock.png"/> <a  type="item" element="'+element+'" class="sharingLink">'+json.browse.share+'</a></li>';
         html+='<li class="removeItemLi"><img alt="" src="'+json.global.coreWebroot+'/public/images/icons/close.png"/> <a onclick="removeItem('+element+');">'+json.browse['removeItem']+'</a></li>';
